requirements.txt
Flask==2.0.3
WTForms==3.0.0

Pillow~=10.0.1
python-dotenv~=1.0.0